+++ examples/vanilla-ts-esm/src/mux-player-classic-theme.ts
@@ -1,0 +1,1 @@
+export * as ThemeClassic from "@mux/mux-player/themes/classic";