+++ examples/vanilla-ts-esm/src/mux-player-gerwig-theme.ts
@@ -1,1 +1,0 @@
-export * as ThemeGerwig from "@mux/mux-player/themes/gerwig";