+++ packages/mux-player-react/src/index.tsx
@@ -1,6 +1,14 @@
 import React, { useEffect } from 'react';
 import type { CSSProperties } from 'react';
-import type { StreamTypes, PlaybackTypes, CmcdTypes } from '@mux/playback-core';
+import type {
+  StreamTypes,
+  PlaybackTypes,
+  CmcdTypes,
+  MaxResolutionValue,
+  MinResolutionValue,
+  RenditionOrderValue,
+} from '@mux/playback-core';
+import { MaxResolution, MinResolution, RenditionOrder } from '@mux/playback-core';
 import { MediaError } from '@mux/mux-player';
 import type MuxPlayerElement from '@mux/mux-player';
 import type { Tokens, MuxPlayerElementEventMap } from '@mux/mux-player';
@@ -10,7 +18,7 @@ import { useCombinedRefs } from './useCombinedRefs';
 import useObjectPropEffect, { defaultHasChanged } from './useObjectPropEffect';
 import { getPlayerVersion } from './env';
 
-export { MediaError };
+export { MediaError, MaxResolution, MinResolution, RenditionOrder };
 
 type ValueOf<T> = T[keyof T];
 interface GenericEventListener<T extends Event = CustomEvent> {
@@ -41,8 +49,11 @@ type MuxMediaPropTypes = {
   // debug: Options["debug"] & Hls["config"]["debug"];
   debug: boolean;
   disableCookies: boolean;
+  disablePictureInPicture?: boolean;
   // metadata: Partial<Options["data"]>;
   metadata: { [k: string]: any };
+  extraSourceParams: Record<string, any>;
+  _hlsConfig: MuxPlayerElement['_hlsConfig'];
   beaconCollectionDomain: string;
   customDomain: string;
   playbackId: string;
@@ -66,7 +77,9 @@ export type MuxPlayerProps = {
   playerSoftwareName?: string;
   forwardSeekOffset?: number;
   backwardSeekOffset?: number;
-  maxResolution?: string;
+  maxResolution?: MaxResolutionValue;
+  minResolution?: MinResolutionValue;
+  renditionOrder?: RenditionOrderValue;
   metadataVideoId?: string;
   metadataVideoTitle?: string;
   metadataViewerUserId?: string;
@@ -170,10 +183,14 @@ const usePlayer = (
     playbackRates,
     currentTime,
     themeProps,
+    extraSourceParams,
+    _hlsConfig,
     ...remainingProps
   } = props;
   useObjectPropEffect('playbackRates', playbackRates, ref);
   useObjectPropEffect('metadata', metadata, ref);
+  useObjectPropEffect('extraSourceParams', extraSourceParams, ref);
+  useObjectPropEffect('_hlsConfig', _hlsConfig, ref);
   useObjectPropEffect('themeProps', themeProps, ref);
   useObjectPropEffect('tokens', tokens, ref);
   useObjectPropEffect('playbackId', playbackId, ref);