+++ packages/mux-video-react/src/index.tsx
@@ -24,16 +24,16 @@ const playerSoftwareVersion = getPlayerVersion();
 const playerSoftwareName = 'mux-video-react';
 
 const MuxVideo = React.forwardRef<HTMLVideoElement | undefined, Partial<Props>>((props, ref) => {
-  const { playbackId, src: outerSrc, children, autoPlay, preload, streamType, disableCookies, ...restProps } = props;
+  const { playbackId, src: outerSrc, children, autoPlay, preload, streamType, disableCookies, _hlsConfig, ...restProps } = props;
 
   const [playerInitTime] = useState(generatePlayerInitTime());
-  const [src, setSrc] = useState<MuxMediaProps['src']>(toMuxVideoURL(playbackId) ?? outerSrc);
+  const [src, setSrc] = useState<MuxMediaProps['src']>(toMuxVideoURL(props) ?? outerSrc);
   const playbackCoreRef = useRef<PlaybackCore | undefined>(undefined);
   const innerMediaElRef = useRef<HTMLVideoElement>(null);
   const mediaElRef = useCombinedRefs(innerMediaElRef, ref);
 
   useEffect(() => {
-    setSrc(toMuxVideoURL(playbackId) ?? outerSrc);
+    setSrc(toMuxVideoURL(props) ?? outerSrc);
   }, [outerSrc, playbackId]);
 
   useEffect(() => {
@@ -82,6 +82,7 @@ MuxVideo.propTypes = {
   // Improve this by adding a full shape() definition for all metadata props
   // metadata: PropTypes.shape({}),
   metadata: PropTypes.any,
+  _hlsConfig: PropTypes.any,
   beaconCollectionDomain: PropTypes.string,
   playbackId: PropTypes.string,
   playerInitTime: PropTypes.number,