+++ packages/playback-core/src/index.ts
@@ -26,18 +26,16 @@ import {
   toTargetLiveWindowFromPlaylistType,
   addEventListenerWithTeardown,
 } from './util';
-import {
-  StreamTypes,
-  PlaybackTypes,
-  ExtensionMimeTypeMap,
-  CmcdTypes,
-  type ValueOf,
-  type PlaybackCore,
-  type MuxMediaProps,
-  type MuxMediaPropsInternal,
-  HlsPlaylistTypes,
-  MediaTypes,
+import type {
+  ValueOf,
+  PlaybackCore,
+  MuxMediaProps,
+  MuxMediaPropsInternal,
+  MaxResolutionValue,
+  MinResolutionValue,
+  RenditionOrderValue,
 } from './types';
+import { StreamTypes, PlaybackTypes, ExtensionMimeTypeMap, CmcdTypes, HlsPlaylistTypes, MediaTypes } from './types';
 export {
   mux,
   Hls,
@@ -230,12 +228,66 @@ export const generatePlayerInitTime = () => {
 
 export const generateUUID = mux.utils.generateUUID;
 
-export const toMuxVideoURL = (playbackId?: string, { domain = MUX_VIDEO_DOMAIN, maxResolution = '' } = {}) => {
-  if (!playbackId) return undefined;
-  const [idPart, queryPart = ''] = toPlaybackIdParts(playbackId);
-  const url = new URL(`https://stream.${domain}/${idPart}.m3u8${queryPart}`);
-  if (maxResolution) {
-    url.searchParams.set('max_resolution', maxResolution);
+type MuxVideoURLProps = Partial<{
+  playbackId: string;
+  customDomain: string;
+  maxResolution: MaxResolutionValue;
+  minResolution: MinResolutionValue;
+  renditionOrder: RenditionOrderValue;
+  tokens: Partial<{
+    playback: string;
+    storyboard: string;
+    thumbnail: string;
+  }>;
+  extraSourceParams: Record<string, any>;
+}>;
+
+export const toMuxVideoURL = ({
+  playbackId: playbackIdWithParams,
+  customDomain: domain = MUX_VIDEO_DOMAIN,
+  maxResolution,
+  minResolution,
+  renditionOrder,
+  tokens: { playback: token } = {},
+  extraSourceParams = {},
+}: MuxVideoURLProps = {}) => {
+  if (!playbackIdWithParams) return undefined;
+  const [playbackId, queryPart = ''] = toPlaybackIdParts(playbackIdWithParams);
+  const url = new URL(`https://stream.${domain}/${playbackId}.m3u8${queryPart}`);
+  /*
+   * All identified query params here can only be added to public
+   * playback IDs. In order to use these features with signed URLs
+   * the query param must be added to the signing token.
+   *
+   * */
+  if (token || url.searchParams.has('token')) {
+    url.searchParams.forEach((_, key) => {
+      if (key != 'token') url.searchParams.delete(key);
+    });
+    if (token) url.searchParams.set('token', token);
+  } else {
+    if (maxResolution) {
+      url.searchParams.set('max_resolution', maxResolution);
+    }
+    if (minResolution) {
+      url.searchParams.set('min_resolution', minResolution);
+      if (maxResolution && +maxResolution.slice(0, -1) < +minResolution.slice(0, -1)) {
+        console.error(
+          'minResolution must be <= maxResolution',
+          'minResolution',
+          minResolution,
+          'maxResolution',
+          maxResolution
+        );
+      }
+    }
+    if (renditionOrder) {
+      url.searchParams.set('rendition_order', renditionOrder);
+    }
+    Object.entries(extraSourceParams).forEach(([k, v]) => {
+      if (v == undefined) return;
+      url.searchParams.set(k, v);
+    });
   }
   return url.toString();
 };
@@ -372,11 +424,14 @@ function useNative(
 
 export const setupHls = (
   props: Partial<
-    Pick<MuxMediaPropsInternal, 'debug' | 'streamType' | 'type' | 'startTime' | 'metadata' | 'preferCmcd'>
+    Pick<
+      MuxMediaPropsInternal,
+      'debug' | 'streamType' | 'type' | 'startTime' | 'metadata' | 'preferCmcd' | '_hlsConfig'
+    >
   >,
   mediaEl: Pick<HTMLMediaElement, 'canPlayType'>
 ) => {
-  const { debug, streamType, startTime: startPosition = -1, metadata, preferCmcd } = props;
+  const { debug, streamType, startTime: startPosition = -1, metadata, preferCmcd, _hlsConfig = {} } = props;
   const type = getType(props);
   const hlsType = type === ExtensionMimeTypeMap.M3U8;
   const shouldUseNative = useNative(props, mediaEl);
@@ -406,8 +461,20 @@ export const setupHls = (
       debug,
       startPosition,
       cmcd,
+      xhrSetup: (xhr, url) => {
+        if (preferCmcd && preferCmcd !== CmcdTypes.QUERY) return;
+        const urlObj = new URL(url);
+        if (!urlObj.searchParams.has('CMCD')) return;
+        const cmcdVal = (urlObj.searchParams.get('CMCD')?.split(',') ?? [])
+          .filter((cmcdKVStr) => cmcdKVStr.startsWith('sid') || cmcdKVStr.startsWith('cid'))
+          .join(',');
+        urlObj.searchParams.set('CMCD', cmcdVal);
+
+        xhr.open('GET', urlObj);
+      },
       ...defaultConfig,
       ...streamTypeConfig,
+      ..._hlsConfig,
     }) as HlsInterface;
 
     return hls;